+++ src/main/java/br/com/integratefarma/funcionario/repository/FuncionarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, Long> {
+public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, Integer> {
 
 }
