+++ src/main/java/br/com/integratefarma/itemvenda/repository/ItemVendaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemVendaRepository extends JpaRepository<ItemVendaEntity, Long> {
+public interface ItemVendaRepository extends JpaRepository<ItemVendaEntity, Integer> {
 
 }
