+++ src/main/java/br/com/integratefarma/venda/repository/VendaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VendaRepository extends JpaRepository<VendaEntity, Long> {
+public interface VendaRepository extends JpaRepository<VendaEntity, Integer> {
 
 }
